Move all background code to /backgrounds/
