+++ services/src/backgrounds/unsplash/types.ts
@@ -1,26 +1,8 @@
-export type Provider = 'pixabay' | 'unsplash'
-
-/**
- * Unified schema returned for Bonjourr Images
- */
-export interface Image {
-	/* All providers */
-	format: 'image'
-	page: string
-	username: string
-	urls: {
-		full: string
-		medium: string
-		small: string
-	}
-
-	/* Unsplash only */
-	color?: string
-	name?: string
-	city?: string
-	country?: string
-	download?: string
-	exif?: {
+export interface UnsplashPhoto {
+	color: string
+	blur_hash: string
+	description: string
+	exif: {
 		make: string
 		model: string
 		exposure_time: string
@@ -28,21 +10,23 @@ export interface Image {
 		focal_length: string
 		iso: number
 	}
-}
-
-/**
- * Unified schema returned for Bonjourr Videos
- */
-export interface Video {
-	format: 'video'
-	page: string
-	username: string
-	duration: number
-	thumbnail: string
+	location: {
+		name: string
+		city: string
+		country: string
+	}
 	urls: {
-		full: string
-		medium: string
-		small: string
+		raw: string
+	}
+	links: {
+		html: string
+	}
+	user: {
+		username: string
+		name: string
+		links: {
+			html: string
+		}
 	}
 }
